+++ src/QueryBuilder/Query/BuilderPrototypeQuery.php
@@ -8,8 +8,13 @@ use Solarium\QueryType\Select\Query\Query;
 
 final class BuilderPrototypeQuery extends AbstractSolrSelectQuery
 {
-    /** @param ApplicatorInterface[] $applicators */
-    public function __construct(private array $applicators)
+    public const METADATA_KEY_ENTITY = 'entity';
+
+    /**
+     * @param ApplicatorInterface[] $applicators
+     * @param array<string, string> $metadata
+     */
+    public function __construct(private array $applicators, private array $metadata = [])
     {
     }
 
@@ -20,5 +25,28 @@ final class BuilderPrototypeQuery extends AbstractSolrSelectQuery
         }
 
         return $select;
+    }
+
+    public function setMetadata(string $key, string $value): void
+    {
+        $this->metadata[$key] = $value;
+    }
+
+    public function getMetadata(string $key): ?string
+    {
+        return $this->metadata[$key] ?? null;
+    }
+
+    public function getProfilerData(): ?array
+    {
+        $data = [
+            'Applicators' => array_map(get_class(...), $this->applicators),
+        ];
+
+        if (!empty($this->metadata)) {
+            $data['Metadata'] = $this->metadata;
+        }
+
+        return parent::getProfilerData() + $data;
     }
 }